Refactor Printer to simplify use
Remove the need to handle path dependent tasty.Reflection
