+++ library/src/scala/tasty/reflect/SettingsOps.scala
@@ -3,10 +3,10 @@ package scala.tasty.reflect
 trait SettingsOps extends ReflectionCore {
 
   /** Compiler settings */
-  def settings(implicit ctx: Context): Settings
+  def settings: Settings
 
   trait SettingsAPI {
-    def color(implicit ctx: Context): Boolean
+    def color: Boolean
   }
   implicit def SettingsDeco(settings: Settings): SettingsAPI
